+++ src/test/java/net/minestom/server/ServerProcessTest.java
@@ -17,7 +17,10 @@ public class ServerProcessTest {
         assumeTrue(System.getenv("GITHUB_ACTIONS") == null);
 
         AtomicReference<ServerProcess> process = new AtomicReference<>();
-        assertDoesNotThrow(() -> process.set(MinecraftServer.updateProcess()));
+        assertDoesNotThrow(() -> {
+            MinecraftServer.init();
+            process.set(MinecraftServer.process());
+        });
         assertDoesNotThrow(() -> process.get().start(new InetSocketAddress("localhost", 25565)));
         assertThrows(Exception.class, () -> process.get().start(new InetSocketAddress("localhost", 25566)));
         assertDoesNotThrow(() -> process.get().stop());
@@ -28,7 +31,8 @@ public class ServerProcessTest {
         // These like to fail on github actions
         assumeTrue(System.getenv("GITHUB_ACTIONS") == null);
 
-        var process = MinecraftServer.updateProcess();
+        MinecraftServer.init();
+        var process = MinecraftServer.process();
         process.start(new InetSocketAddress("localhost", 25565));
         var ticker = process.ticker();
         assertDoesNotThrow(() -> ticker.tick(System.currentTimeMillis()));